requirements.txt
autogen_agentchat==0.2.37
chromadb==0.5.18
google_search_results==2.4.2
gradio==5.6.0
gradio_client==1.4.3
ipython==8.12.3
llama_index==0.11.23
nltk==3.9.1
numpy==2.1.3
openai==1.54.4
Pillow==11.0.0
setuptools==75.1.0
torch==2.4.0+cu118
tqdm==4.66.5
